Try to correct issue with rmdir on windows

+++ smithy-cli/src/it/java/software/amazon/smithy/cli/IntegUtils.java
@@ -150,6 +150,8 @@ public final class IntegUtils {
     }
 
     public static void clearCacheDirIfExists() {
-        IoUtils.rmdir(SMITHY_ROOT_CACHE_PATH);
+        if (Files.exists(SMITHY_ROOT_CACHE_PATH)) {
+            IoUtils.rmdir(SMITHY_ROOT_CACHE_PATH);
+        }
     }
 }
